% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_l.R
\name{ch_l}
\alias{ch_l}
\title{Function for optimisation of spending allocation}
\usage{
ch_l(p)
}
\arguments{
\item{p}{strength vector}
}
\value{
strength info
}
\description{
Function for optimisation of spending allocation
}
\examples{
ch_l(c(5 ,5, 5)

}
