% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{data for dummy jogo do pau tournament}
\format{
A data frame with 16 rows
Each row corresponds to one player
}
\source{
{in-house}
}
\usage{
dummy_data
}
\description{
data for dummy jogo do pau tournament
}
\examples{
print(dummy_data)


}
\keyword{datasets}
