% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formos_l.R
\name{formos_l}
\alias{formos_l}
\title{Truncated linear formos}
\usage{
formos_l(x, k = 1.8)
}
\arguments{
\item{x}{x}

\item{k}{scale}
}
\value{
formos map
}
\description{
Truncated linear map
}
\examples{
formos_l(1:20, 2)

}
