% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdp_data.R
\docType{data}
\name{jdp_data}
\alias{jdp_data}
\title{data for jogo do pau tournament}
\format{
A data frame with 100 rows
Each row corresponds to one simulated match
}
\source{
{in-house}
}
\usage{
jdp_data
}
\description{
data for jogo do pau tournament
}
\examples{
print(jdp_data)


}
\keyword{datasets}
