% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgame_play_l.R
\name{mgame_play_l}
\alias{mgame_play_l}
\title{Simulate multiple matches}
\usage{
mgame_play_l(n, a1_m, a2_m, seed = NULL)
}
\arguments{
\item{n}{number of matches}

\item{a1_m}{matrix of allocations for player 1}

\item{a2_m}{matrix of allocations for player 2}

\item{seed}{seed}
}
\value{
simulated match info
}
\description{
Simulate multiple matches
}
\examples{
mgame_play_l(16, dummy_data[, 2:4], dummy_data[16:1, 2:4], seed = 15)

}
