% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgame_sim_l.R
\name{mgame_sim_l}
\alias{mgame_sim_l}
\title{Simulate multiple matches}
\usage{
mgame_sim_l(n, seed = NULL)
}
\arguments{
\item{n}{number of matches}

\item{seed}{seed}
}
\value{
simulated match info
}
\description{
Simulate multiple matches
}
\examples{
mgame_sim_l(100, seed = 15)

}
