% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarilhos_l.R
\name{sarilhos_l}
\alias{sarilhos_l}
\title{Truncated linear sarilhos}
\usage{
sarilhos_l(x, k = 1.2)
}
\arguments{
\item{x}{x}

\item{k}{scale}
}
\value{
sarilhos map
}
\description{
Truncated linear map
}
\examples{
sarilhos_l(1:20, 2)

}
