% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_l.R
\name{series_l}
\alias{series_l}
\title{Truncated linear series}
\usage{
series_l(x, k = 1.5)
}
\arguments{
\item{x}{x}

\item{k}{scale}
}
\value{
series map
}
\description{
Truncated linear map
}
\examples{
series_l(1:20, 2)

}
