% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strengths_l.R
\name{strengths_l}
\alias{strengths_l}
\title{Strengths}
\usage{
strengths_l(p1, p2)
}
\arguments{
\item{p1}{team 1 allocation}

\item{p2}{team 2 allocation}
}
\value{
strength info
}
\description{
Strengths
}
\examples{
strengths_l(c(5, 5, 5), c(2, 9, 4))

}
